requirements copy.txt
fastapi==0.110.0
yt-dlp==2024.04.09
openai==1.42.0
uvicorn==0.29.0
python-dotenv
chainlit
